big structure change, to improve organization and test support *pending removal of commom package* *pending removal of db connection from gin context *
