--- model/tag_test.go
@@ -1,12 +1,0 @@
-package model
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-)
-
-func TestTag_TableName(t *testing.T) {
-	tags := Tag{}
-	assert.Equal(t, "public.tags", tags.TableName())
-}
