+++ pkg/domain/base_test.go
@@ -1,0 +1,13 @@
+package domain
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestBaseModel_IsNew(t *testing.T) {
+	new := BaseModel{}
+
+	assert.True(t, new.IsNew())
+}
