+++ pkg/domain/tag_test.go
@@ -1,0 +1,12 @@
+package domain
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestTag_TableName(t *testing.T) {
+	tags := Tag{}
+	assert.Equal(t, "public.tags", tags.TableName())
+}
