+++ pkg/http/auth/types.go
