--- url/index_test.go
@@ -1,21 +1,0 @@
-package url
-
-import (
-	"net/http"
-	"net/http/httptest"
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-)
-
-func TestIndexHandler(t *testing.T) {
-	// router = gin.New()
-	router.GET("/", IndexHandler)
-
-	w := httptest.NewRecorder()
-	req, _ := http.NewRequest("GET", "/", nil)
-	router.ServeHTTP(w, req)
-
-	assert.Equal(t, http.StatusOK, w.Code)
-	assert.Contains(t, w.Body.String(), "\"status\":\"ok\"")
-}
