--- url/main_test.go
@@ -1,14 +1,0 @@
-package url
-
-import (
-	"os"
-	"testing"
-
-	"github.com/gin-gonic/gin"
-)
-
-var router = gin.Default()
-
-func TestMain(m *testing.M) {
-	os.Exit(m.Run())
-}
